fix to progress/congestion map alignment
